+++ src/app.h
@@ -2,6 +2,7 @@
 
 #include "keyboard_hook.h"
 #include "ime_controller.h"
+#include "settings.h"
 #include "tray_icon.h"
 
 #include <windows.h>
@@ -25,4 +26,5 @@ class App {
   TrayIcon tray_icon_;
   KeyboardHook keyboard_hook_;
   ImeController ime_controller_;
+  AppSettings settings_;
 };